split environment.py into multiple modules
